purchase test

+++ app/Services/Order/PurchaseService.php
@@ -20,10 +20,10 @@ class PurchaseService
     /**
      * get maximum amount that user can but
      * @param mixed $platform
-     * @param Coin|array|int $coin
+     * @param Coin|array|float $coin
      * @return float
      */
-    public static function maximumAvailableAmount(mixed $platform , Coin|array|int $coin) : float
+    public static function maximumAvailableAmount(mixed $platform , Coin|array|float $coin) : float
     {
         $platform = PlatformService::find($platform);
 
@@ -32,7 +32,7 @@ class PurchaseService
         elseif ( is_array($coin) and isset($coin['price']))
             $coin = $coin['price'] ;
         else
-            $coin = (int) $coin ;
+            $coin = (float) $coin ;
         $availableAmount = $platform->available_tether / $coin;
         $mult = pow(10, 8);
         return floor($availableAmount * $mult) / $mult;
@@ -40,18 +40,18 @@ class PurchaseService
 
     /**
      * generate order price
-     * @param Coin|array|int $coin
+     * @param Coin|array|float $coin
      * @param float $amount
      * @return float
      */
-    public static function price(Coin|array|int $coin , float $amount) : float
+    public static function price(Coin|array|float $coin , float $amount) : float
     {
         if ( $coin instanceof Coin)
             $coin = $coin->price ;
         elseif ( is_array($coin) and isset($coin['price']))
             $coin = $coin['price'] ;
         else
-            $coin = (int) $coin ;
+            $coin = (float) $coin ;
 
         return $amount * $coin;
     }
